Reorganize scenario-based test helpers
