+++ airbyte-cdk/python/unit_tests/sources/scenario_based/helpers.py
